ban sua moi nhat

+++ ajax/gt_xoa.php
@@ -1,7 +1,10 @@
 <?
     include("config.php");
-    $id        = $_POST['id'];
+    $id = getValue('id', 'int', 'POST', '');
 
-     
-    $delete_gt = new db_query("DELETE FROM `ds_gia_tri_dg` WHERE `id` = '$id' ");
+    if($id != ""){
 
+        $delete_gt = new db_query("DELETE FROM `ds_gia_tri_dg` WHERE `id` = '$id' ");
+    }else{
+        echo "Bạn xóa tiêu chí đánh giá không thành công, vui lòng thử lại";
+    }
